スキルポイントを小数点対応

+++ src/main/scala/com/ryoserver/SkillSystems/SkillPoint/SkillPointData.scala
@@ -6,18 +6,18 @@ import org.bukkit.entity.Player
 
 class SkillPointData(ryoServerAssist: RyoServerAssist) {
 
-  def getSkillPoint(p: Player): Int = {
+  def getSkillPoint(p: Player): Double = {
     val sql = new SQL(ryoServerAssist)
     val rs = sql.executeQuery(s"SELECT SkillPoint FROM Players WHERE UUID='${p.getUniqueId.toString}'")
-    var sp = 0
+    var sp = 0.0
     if (rs.next()) {
-      sp = rs.getInt("SkillPoint")
+      sp = rs.getDouble("SkillPoint")
     }
     sql.close()
     sp
   }
 
-  def setSkillPoint(p: Player, point: Int): Unit = {
+  def setSkillPoint(p: Player, point: Double): Unit = {
     val sql = new SQL(ryoServerAssist)
     sql.executeSQL(s"UPDATE Players SET SkillPoint='${point}' WHERE UUID='${p.getUniqueId.toString}';")
     sql.close()
